Data Structures and Algorithms 3.2
Palindrome is a word, phrase, number, or other sequence of characters
which reads the same backward or forward. Write a function that checks
if a given string of one or more words is a palindrome or not. Your
function can assume that each word consists of at least two letters and
that each word is separated by one space, with no punctuation marks. Use
the following palindromes (e.g. “Never odd or even”, “Hannah”, “If I had
a hifi”) as well as any non-palindrome strings to check your code: